% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sample_res.R
\name{fetch_sample_res}
\alias{fetch_sample_res}
\title{"fetch: sample results"}
\usage{
fetch_sample_res(df, site_notation, dets, start_date, end_date, limit)
}
\arguments{
\item{df}{if the sample point function is saved to a dataframe the column
can be used (df$notation)}

\item{site_notation}{Either a single site can be submitted ("site_notation1"),
a number or specifc sites (c(site_notation1,site_notation2))}

\item{dets}{one or more determinands can be requested using the same logic
as above}

\item{start_date}{a date for the start of the data in a YYYY-mm-dd format}

\item{end_date}{a date for the end of the data in a YYYY-mm-dd format}

\item{limit}{if you want to run the query quickly you can set a numeric limit}
}
\value{
returns the output of the API query based on the user search
}
\description{
"Pulls the data from the DDSP API for the required sample points acquired
from the sample_points function"
}
\examples{
fetch_sample_res(site_notation = "AN-FBRO100F", dets = c("0085","0111"))

# will return a dataframe containing the BOD (0085) and ammonia (0111)
# for all the sample point provided

}
