% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_dets.R
\name{helper_dets}
\alias{helper_dets}
\title{"helper: determinands"}
\usage{
helper_dets(det_name)
}
\arguments{
\item{det_name}{word or part of word to search the list for (ignores case)}
}
\value{
returns the output of the API query based on the user search
}
\description{
"Searches the determinands (dets) list for potential determinands of interest based
on pattern matches"
}
\examples{
helper_dets("BOD")

# will provide a df with all the dets that contain "bod"

}
