% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_ea_area.R
\name{helper_ea_area}
\alias{helper_ea_area}
\title{"helper: EA Areas"}
\usage{
helper_ea_area(ea_area)
}
\arguments{
\item{ea_area}{word or part of word to search the list for (ignores case).
if left blank returns all EA areas.}
}
\value{
returns the output of the API query based on the user search
}
\description{
"Searches the area list for areas of interest based
on pattern matches"
}
\examples{
smpt_types <- helper_ea_areas(ea_area = "man")

# will provide a df with all the sample point types that contain "mine"

}
